Initial
This is the initial and final commit to this project. This project was a simple beacon script to help me understand more about the socket library by documenting and researching it a bit. It has much more depth than in this project alone but it helped me to understand how it works in much more depth.